changed how blinking request is sent + same pattern shows up on phone
in order to ask a router to blink, you need to click and hold it on the list.
the pattern will show up on the phone as well.
and a toast message will tell you that a blink request was sent.
it's in a new separate activity called ThirdScenario.
(to try it just change the next button to take you to ThirdScenario instead of FirstScenario)